pyproject.toml
[project]
name = "authorizationfastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "clasterpack",email = "clasterpack@gmail.com"}
]
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "dishka (>=1.6.0,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "cryptography (>=45.0.6,<46.0.0)",
    "orjson (>=3.11.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"

